Moved Kafka-specific classes out of client.ext package
The IdStrategy is Kakfa-specific, and DocumentWriteOperationBuilder can be used for inspiration for something reusable in the Java Client.
