+++ src/main/java/com/marklogic/kafka/connect/sink/idstrategy/DefaultStrategy.java
@@ -1,12 +1,15 @@
-package com.marklogic.client.id.strategy;
+package com.marklogic.kafka.connect.sink.idstrategy;
 
 import com.marklogic.client.io.marker.AbstractWriteHandle;
 
-public class KafkaMetaStrategy implements IdStrategy {
+import java.util.UUID;
+
+
+public class DefaultStrategy implements IdStrategy {
 
     @Override
     public String generateId(AbstractWriteHandle content, String topic, Integer partition, long offset) {
-        return topic + "/" + partition.toString() + "/" + offset;
+        return UUID.randomUUID().toString();
     }
 
 }
