+++ src/main/java/com/marklogic/kafka/connect/sink/idstrategy/HashedJSONPathsStrategy.java
@@ -1,4 +1,4 @@
-package com.marklogic.client.id.strategy;
+package com.marklogic.kafka.connect.sink.idstrategy;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
