+++ src/main/java/com/marklogic/kafka/connect/sink/idstrategy/IdStrategyFactory.java
@@ -1,4 +1,4 @@
-package com.marklogic.client.id.strategy;
+package com.marklogic.kafka.connect.sink.idstrategy;
 
 import com.marklogic.kafka.connect.sink.MarkLogicSinkConfig;
