+++ src/main/java/com/marklogic/kafka/connect/sink/idstrategy/KafkaMetaStrategy.java
@@ -1,15 +1,12 @@
-package com.marklogic.client.id.strategy;
+package com.marklogic.kafka.connect.sink.idstrategy;
 
 import com.marklogic.client.io.marker.AbstractWriteHandle;
 
-import java.util.UUID;
-
-
-public class DefaultStrategy implements IdStrategy {
+public class KafkaMetaStrategy implements IdStrategy {
 
     @Override
     public String generateId(AbstractWriteHandle content, String topic, Integer partition, long offset) {
-        return UUID.randomUUID().toString();
+        return topic + "/" + partition.toString() + "/" + offset;
     }
 
 }
