+++ src/main/java/com/marklogic/kafka/connect/source/DocumentWriteOperationBuilder.java
@@ -1,4 +1,4 @@
-package com.marklogic.client.ext.document;
+package com.marklogic.kafka.connect.source;
 
 import com.marklogic.client.document.DocumentWriteOperation;
 import com.marklogic.client.ext.util.DefaultDocumentPermissionsParser;
